perf-test/requirements.txt
locust==2.24.1
requests==2.31.0
python-dotenv==1.0.0
faker==23.0.0
pytest==7.4.3
